rename app / dashboard components based on pr feedback
