+++ ui/dashboards/src/components/PersesDashboard.tsx
@@ -11,16 +11,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { Box } from '@mui/material';
+import { Box, BoxProps } from '@mui/material';
+import { DashboardResource } from '@perses-dev/core';
 import { combineSx } from '@perses-dev/components';
 import { useState } from 'react';
 import { VariableList, Dashboard } from '../components';
 import AddPanel from '../components/AddPanel/AddPanel';
 import { DashboardToolbar } from '../components/DashboardToolbar';
 import { useDashboard } from '../context';
-import { ViewDashboardProps } from './ViewDashboard';
 
-export const DashboardApp = (props: ViewDashboardProps) => {
+export interface PersesDashboardProps extends BoxProps {
+  dashboardResource: DashboardResource;
+}
+
+export const PersesDashboard = (props: PersesDashboardProps) => {
   const { dashboardResource, sx, children, ...others } = props;
   const [showAddPanel, setShowAddPanel] = useState(false);
   const { dashboard } = useDashboard();
